pyproject.toml
[project]
name = "computer-use-agent"
version = "0.1.0"
description = "A computer use agent using Claude with human-in-the-loop capabilities"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "pillow>=10.0.0",
    "docker>=7.0.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "capmonstercloudclient",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.5.0",
]

[project.scripts]
computer-use-agent = "computer_use_agent.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/computer_use_agent"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
